Properly expose the Eyeglass class as having a functional 'newless' constructor.
